+++ ui/user/src/lib/components/admin/audit-log-exports/CreateScheduleForm.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import { page } from '$app/state';
 	import Select from '$lib/components/Select.svelte';
+	import { filterVisibleExportFields } from '$lib/components/admin/audit-log-exports/filterFields';
 	import Loading from '$lib/icons/Loading.svelte';
 	import {
 		type LLMAuditLogURLFilters,
@@ -49,7 +50,7 @@
 			timezone: Intl.DateTimeFormat().resolvedOptions().timeZone
 		},
 		retentionPeriodInDays: 30,
-		sourceTypes: ['mcp'] as string[],
+		sourceTypes: ['mcp'],
 		filters: {
 			user_id: '',
 			mcp_id: '',
@@ -628,23 +629,12 @@
 		'device_id'
 	]);
 	const visibleScheduleFilterRows = $derived(
-		auditScheduleAdvancedFilterRows.filter((row) => {
-			const hasMCPFilters = [...mcpScheduleFilterKeys].some(
-				(key) => form.filters[key as keyof typeof form.filters]
-			);
-			const hasLocalFilters = [...localScheduleFilterKeys].some(
-				(key) => form.filters[key as keyof typeof form.filters]
-			);
-			// Source-specific filters require exactly one selected source (the backend rejects them for
-			// mixed-source schedules). Keep a group visible if it still holds values so the user can
-			// clear a stale selection after switching sources, instead of both groups hiding at once.
-			const onlyMCP = form.sourceTypes.length === 1 && form.sourceTypes.includes('mcp');
-			const onlyLocal =
-				form.sourceTypes.length === 1 && form.sourceTypes.includes('local_agent_tool_call');
-			if (mcpScheduleFilterKeys.has(row.filterKey)) return hasMCPFilters || onlyMCP;
-			if (localScheduleFilterKeys.has(row.filterKey)) return hasLocalFilters || onlyLocal;
-			return true;
-		})
+		filterVisibleExportFields(
+			form,
+			auditScheduleAdvancedFilterRows,
+			mcpScheduleFilterKeys,
+			localScheduleFilterKeys
+		)
 	);
 
 	function normalizeSourceTypes(sourceTypes: string[] | undefined): string[] {
